feat: restructing of codes
