+++ src/tailwind.js
@@ -40,6 +40,7 @@ const localConfig = {
         ],
       },
       minWidth: {
+        46: "11.5rem",
         135: "135px",
         200: "200px",
         xxs: "15.5rem",
